系统消息代码优化

+++ src/main/java/com/bc/pmpheep/back/service/CmsContentQuartzJob.java
@@ -8,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 
 import com.bc.pmpheep.back.po.CmsContent;
 import com.bc.pmpheep.back.po.CmsSchedule;
+import com.bc.pmpheep.back.util.CollectionUtil;
 import com.bc.pmpheep.back.util.DateUtil;
 import com.bc.pmpheep.service.exception.CheckedServiceException;
 
@@ -39,17 +40,19 @@ public class CmsContentQuartzJob {
         // 获取当前时间前5分钟定时任务数
         List<CmsSchedule> cmsSchedules =
         cmsScheduleService.getCmsScheduleList(new CmsSchedule(DateUtil.getCurrentTime()));
-        List<Long> contentIds = new ArrayList<Long>(cmsSchedules.size());// contentId集合
-        List<CmsContent> cmsContents = new ArrayList<CmsContent>(cmsSchedules.size());// content对象集合
-        for (CmsSchedule cmsSchedule : cmsSchedules) {
-            contentIds.add(cmsSchedule.getContentId());
-            cmsContents.add(new CmsContent(cmsSchedule.getContentId(), true,
-                                           cmsSchedule.getScheduledTime()));
+        if (CollectionUtil.isNotEmpty(cmsSchedules)) {
+            List<Long> contentIds = new ArrayList<Long>(cmsSchedules.size());// contentId集合
+            List<CmsContent> cmsContents = new ArrayList<CmsContent>(cmsSchedules.size());// content对象集合
+            for (CmsSchedule cmsSchedule : cmsSchedules) {
+                contentIds.add(cmsSchedule.getContentId());
+                cmsContents.add(new CmsContent(cmsSchedule.getContentId(), true,
+                                               cmsSchedule.getScheduledTime()));
+            }
+            // 按content对象集合批量更新
+            cmsContentService.updateCmsContentByIds(cmsContents);
+            // 按contentId批量删除
+            cmsScheduleService.deleteCmsScheduleByIds(contentIds);
         }
-        // 按content对象集合批量更新
-        cmsContentService.updateCmsContentByIds(cmsContents);
-        // 按contentId批量删除
-        cmsScheduleService.deleteCmsScheduleByIds(contentIds);
         System.out.println("执行时间" + new Date());
     }
 }
